% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_table.R
\docType{data}
\name{argument_table}
\alias{argument_table}
\title{Argument Table}
\description{
REGEX Match list of possible arguments for requests
}
\author{
Lissa Hyacinth \email{lissahyacinth@gmail.com}
}
\keyword{api}
\keyword{arguments}
\keyword{tfl}
\keyword{unified}
