% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructUrl.R
\name{constructUrl}
\alias{constructUrl}
\title{Format a URL for TfL API}
\usage{
constructUrl(url, app_id, app_key)
}
\arguments{
\item{url}{- Origin URL}

\item{app_id}{TfL App ID}

\item{app_key}{App Key}
}
\description{
\code{constructUrl} returns a formatted URL with AppID and AppKey
}
