% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httrStatusCheck.R
\name{httrStatusCheck}
\alias{httrStatusCheck}
\title{Check HTTR Response against common HTTP Codes}
\usage{
httrStatusCheck(resp)
}
\arguments{
\item{resp}{Response object from \code{httr::GET(url)}}
}
\description{
\code{httrStatusCheck(resp = resp)} returns 0,1, or declares an error.
}
