% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initTFL.R
\name{initTFL}
\alias{initTFL}
\title{Initialise TfL Connection}
\usage{
initTFL(updateAPIKey = FALSE, updateLocalData = FALSE)
}
\arguments{
\item{updateAPIKey}{Boolean - Update stored API Key?}

\item{updateLocalData}{Boolean - Update stored data?}
}
\description{
\code{initTFL} connects the API and creates local tables
}
