% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_table.R
\docType{data}
\name{request_table}
\alias{request_table}
\title{Request Table}
\description{
Documented list of possible requests
}
\author{
Lissa Hyacinth \email{lissahyacinth@gmail.com}
}
\keyword{api}
\keyword{arguments}
\keyword{tfl}
\keyword{unified}
