% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflAPI.R
\name{tflAPI}
\alias{tflAPI}
\title{Query an API with Timeout Handling}
\usage{
tflAPI(formatted_url)
}
\arguments{
\item{formatted_url}{URL for querying an API}
}
\description{
\code{tflAPI(formatted_url = url)} returns a JSON Object or timesout.
}
