% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflRequest.R
\name{tflRequest}
\alias{tflRequest}
\title{Request a JSON object from TfL}
\usage{
tflRequest(request, app_id, app_key, args = list())
}
\arguments{
\item{request}{Written request, e.g "all lines"}

\item{app_id}{TfL App ID}

\item{app_key}{App Key}

\item{args}{2D list of arguments to be supplied to the request, e.g. list("Mode" = "underground")}
}
\description{
\code{tflRequest} returns a JSON object
}
